Fix route admin,vendor and create,edit,delete acara, get acara by user role vendor

+++ app/Http/Controllers/VendorController.php
@@ -9,25 +9,30 @@ use App\Models\Acara;
 use App\Models\Category;
 use Illuminate\Http\Request;
 use App\Notifications\userNotification;
+use Illuminate\Support\Facades\Auth;
 
 class VendorController extends Controller
 {
     public function index()
     {
-        $acaras = Acara::paginate(30);
-        return view('vendor.acara.index', [
+
+        // $acaras = Acara::paginate(30);
+        $user = Auth::user(); // Mendapatkan user yang sedang login
+        $acaras = $user->acara()->paginate(30); // Mengambil acara yang dibuat oleh user/vendor  tersebut
+        return view('vendor-client.acara.index', [
             'acaras' => $acaras
         ]);
     }
 
     public function create()
     {
+
         $categories = Category::all() ?? null;
 
         //dd($categories);
         $acaras = Acara::all();
 
-        return view('vendor.acara.create', [
+        return view('vendor-client.acara.create', [
             'acaras' => $acaras,
             'categories' => $categories,
         ]);
@@ -60,24 +65,27 @@ class VendorController extends Controller
 
 
         $acara = Acara::create($request->except('files'));
-        // $users = User::all();
-        // Notification::send($users, new userNotification($acara));
+        $user = Auth::user();
+        $user->acara()->attach($acara->id);
 
         //return to index
-        return redirect()->route('acara.index')->with('success', 'Acara berhasil di buat ');
+        return redirect()->route('vendor.acara.index')->with('success', 'Acara berhasil di buat ');
     }
     public function edit(Acara $acara)
     {
         $categories = Category::all();
         $acaras = Acara::all();
-        return view('vendor.acara.edit', [
+        return view('vendor-client.acara.edit', [
             'acara' => $acara,
             'categories' => $categories,
         ]);
     }
 
-    public function update(Request $request, string $id)
+    public function update(Request $request, Acara $acara)
     {
+        if (!Auth::user()->acara->contains($acara->id)) {
+            return redirect()->route('vendor-client.acara.index')->with('error', 'Anda tidak memiliki izin untuk mengedit acara ini.');
+        }
         $request->validate([
             'name' => 'required',
             'description' => 'required',
@@ -103,13 +111,17 @@ class VendorController extends Controller
 
 
         //update Acara
-        Acara::find($id)->update($request->except('files'));
-        return redirect()->route('acara.index')->with('success', 'Acara berhasil di Edit');
+        $acara->update($request->except('files'));
+        return redirect()->route('vendor.acara.index')->with('success', 'Acara berhasil di buat ');
     }
     public function destroy(Acara $acara)
     {
+        if (!Auth::user()->acara->contains($acara->id)) {
+            return redirect()->route('vendor-client.acara.index')->with('error', 'Anda tidak memiliki izin untuk menghapus acara ini.');
+        }
+
         $acara->delete();
         // dd($acara);
-        return redirect()->route('acara.index')->with('success', 'Acara berhasil di hapus');
+        return redirect()->route('vendor-client.acara.index')->with('success', 'Acara berhasil di hapus');
     }
 }
